Responsible-Gaming/backend/requirements.txt
Flask==2.1.2
google-api-python-client==2.61.0
PyJWT==2.4.0
